backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
requests==2.28.2
pymongo==4.3.3
python-dotenv==1.0.0
openai==0.27.4
python-dateutil==2.8.2
twython==3.9.1
linkedin-api
pillow==9.5.0
gunicorn==20.1.0
